refactor to utils/
